Make `withProviderClassesFrom` work with Gradle
The `withProviderClassesFrom`-method takes a relative path as an argument, usually `targets/classes`. That path isn't resolved correctly for Gradle projects.
Moreover, the logic assumes that the `META-INF` directory is located in the extension classes folder. This is also not true for Gradle projects.

The issue is demonstrated in this [example test].

This commit contains a possible fix, in which the classes directory is resolved relative to the current user directory and the `META-INF` directory is sought in the location we suspect it with Gradle, and added, if found.

Also added a few unit tests for the path resolution functionality. The tests use the old logic to show that, under default conditions, the outcome didn't change for Maven layouts.

[example test]: https://github.com/objecttrouve/keycloak-testcontainers-gradle-issue-example/blob/main/lib/src/test/java/org/objecttrouve/keycloak/testcontainers/issues/gradleworkingdir/SomeKeycloakSpiTest.java