API: cfg of multiple classes + resources directories
(Non-breaking) API extension that allows the specification of multiple directories for compiled classes and resources (such as the `META-INF` directory).

This makes the logic agnostic to the build tool and flexible enough to work with both Maven and Gradle.

On the other hand, the API gets bloated by having multiple methods for the same configuration goal. Therefore, I suggest to deprecate the methods that take only _one_ classes location path arg and recommend to use the methods that take multiple such paths.

Added some (parameterized) container test cases for the new configuration method.
Left the deprecated config method in place in tests, as it delegates to the new method now.
This way both the old and the new logic are covered.